Move as much progress bar logic out of lib as possible
